requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
supabase==2.0.2
psycopg2-binary==2.9.7
google-generativeai==0.3.2
huggingface_hub==0.20.3
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==0.8.11
reportlab==4.0.4
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3